Rename tests, add smoke test
