+++ .vscode/launch.json
@@ -21,7 +21,8 @@
                 "--runInBand"
             ],
             "console": "integratedTerminal",
-            "internalConsoleOptions": "neverOpen"
+            "internalConsoleOptions": "neverOpen",
+            "preLaunchTask": "yarn: build"
         }
     ]
 }
